WIP - Map as Graph
Trying to build the map as a graph for both graphical representation AND
puzzle representation. Introduced type parameter for relations so one
can either use strings for relations (easy and lw) or some type. Will try
to use sealed classes as per some Medium article for that, could be
cool.